KSK-1 : Moved koski case classes to separate file.
